🔵 reorganize order of functions
